requirements.txt
aniso8601==9.0.1
anyio==3.4.0
ariadne==0.14.0
asgiref==3.4.1
autopep8==1.6.0
cached-property==1.5.2
click==8.0.3
colorama==0.4.4
fastapi==0.71.0
Flask==2.0.2
Flask-GraphQL==2.0.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
graphene==3.0
graphql-core==3.1.7
graphql-relay==3.1.0
graphql-server-core==1.2.0
greenlet==1.1.2
h11==0.12.0
httptools==0.3.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
loguru==0.5.3
MarkupSafe==2.0.1
numpy==1.20.2
pandas==1.2.4
promise==2.3
psycopg2==2.8.6
pycodestyle==2.8.0
pydantic==1.9.0
Pygments==2.11.2
python-dateutil==2.8.1
python-dotenv==0.17.0
python-multipart==0.0.5
pytz==2021.1
PyYAML==6.0
Rx==1.6.1
sentinel==0.3.0
six==1.15.0
sniffio==1.2.0
SQLAlchemy==1.4.29
starlette==0.17.1
strawberry-graphql==0.93.23
toml==0.10.2
typing-extensions==4.0.1
uvicorn==0.16.0
watchgod==0.7
websockets==10.1
Werkzeug==2.0.2
win32-setctime==1.0.3
